+++ CHANGES.rst
@@ -31,7 +31,7 @@ AudioLazy changes history
   - Mock testing for audio output
   - Bugfixes (``envelope.abs``, ``midi2str``, ``StreamTeeHub.blocks``, etc.)
   - Extended domain for some functions by using ``inf`` and ``nan``
-  - Now with 5700+ tests and 80% code coverage
+  - Now with 5700+ tests and 81% code coverage
   - Removed deprecated ``Stream.tee()`` method and DEFAULT_CHUNK_SIZE constant
   - No more distinction between ``__div__`` and ``__truediv__`` (Python 2.7)
   - Now AudioLazy works with Python 3.3!
@@ -70,6 +70,8 @@ AudioLazy changes history
   - Waring-Lagrange polynomial interpolator StrategyDict
   - General resample based on Waring-Lagrange interpolators, working with
     time-varying sample rate
+  - New methods ``Poly.is_polynomial()`` and ``Poly.is_laurent()``
+  - New property ``Poly.order`` for common polynomials
 
 + lazy_stream: